infer ownership of process based on cwd where script is running
Typically scripts create artifacts, and it is inconvenient for these to
be root-owned when running in non-root-owned locations.

Also, there are other places in the npm stack where 'uid' and 'gid'
options may be set, and can cause problems if applied to a spawn() call
when running as non-root.

When running as root, execute scripts using the same uid/gid as the
owning folder.  When running as non-root, do not attempt to change
uid/gid.