Proxy exit signals from foreground processes
This makes it so that the main process will be terminated with the same
signal that a foreground (ie, `stdio: 'inherit'`) process receives.

Thus, if you kick off a long-running script with `npm run`, and then
`^C` to terminate it, the main process will exit with that signal,
instead of treating the signal exit as an error and printing npm's error
banner.

Fixes: https://github.com/npm/cli/issues/2124