fix: handle signals more correctly (#142)
previously we were assuming that npm received the signal first and was
required to forward the signal to child processes, however that seems to
no longer be the case. instead npm and the child process both receive
the signal at the same time. the previous logic has been modified such
that it places a no-op function as the signal handler. this is strictly
to prevent the default behavior of exiting node from happening. once all
child process have exited, the handlers are all removed and we exit
appropriately